# [MPS] Speedup interpolation